Updates and styles VenueManagement component
